[refactor] code formatting
