+++ src/main/java/exceptions/InvalidFilePathException.java
@@ -1,4 +1,4 @@
-package exception;
+package exceptions;
 
 public class InvalidFilePathException extends IllegalArgumentException {
